requirements.txt
Django==1.10.3
django-braces==1.10.0
django-cors-headers==1.3.1
django-filter==1.0.1
django-oauth-toolkit==0.11.0
djangorestframework==3.5.3
gunicorn==19.6.0
httpie==0.9.9
Markdown==2.6.7
nltk==3.2.1
numpy==1.11.3
oauthlib==1.1.2
psycopg2==2.4.5
pycurl==7.19.3
Pygments==2.1.3
requests==2.12.4
scikit-learn==0.18.1
six==1.10.0
uWSGI==2.0.14
virtualenv==15.1.0
whitenoise==3.3.0
dj-database-url==0.4.2
